+++ src/components/layout-and-navigation/BreadCrumb.tsx
@@ -1,8 +1,10 @@
 import Link from 'next/link'
 import clsx from 'clsx'
+import type { ReactNode } from 'react'
+import { Fragment } from 'react'
 
 export type Crumb = {
-  display: string,
+  display: ReactNode,
   link: string,
 }
 
@@ -21,16 +23,21 @@ export const BreadCrumb = ({ crumbs, linkClassName, containerClassName }: BreadC
   const color = 'text-description'
 
   return (
-    <div className={clsx('flex-row-0', containerClassName)}>
-      {crumbs.map((crumb, index) => (
-        <div key={index}>
-          <Link href={crumb.link}
-                className={clsx(linkClassName, { [`${color} hover:brightness-60`]: index !== crumbs.length - 1 })}>
-            {crumb.display}
-          </Link>
-          {index !== crumbs.length - 1 && <span className={clsx(`px-1`, color)}>/</span>}
-        </div>
-      ))}
+    <div className={clsx('flex-row-0.5 items-center', containerClassName)}>
+      {crumbs.map((crumb, index) => {
+        const isLast = index === crumbs.length - 1
+        return (
+          <Fragment key={index}>
+            <Link
+              href={crumb.link}
+              className={clsx('btn-sm hover:bg-description/20', linkClassName, { [color]: !isLast })}
+            >
+              {crumb.display}
+            </Link>
+            {!isLast && <span className={clsx(`px-1`, color)}>/</span>}
+          </Fragment>
+        )
+      })}
     </div>
   )
 }