+++ src/components/layout-and-navigation/Tile.tsx
@@ -3,11 +3,14 @@ import clsx from 'clsx'
 import { Check } from 'lucide-react'
 
 export type TileProps = {
-  title: { value: ReactNode, className?: string },
-  description?: { value: ReactNode, className?: string },
+  title: ReactNode,
+  titleClassName?: string,
+  description?: ReactNode,
+  descriptionClassName?: string,
   onClick?: () => void,
   disabled?: boolean,
   isSelected?: boolean,
+  isListItem?: boolean,
   prefix?: ReactNode,
   suffix?: ReactNode,
   className?: string,
@@ -21,14 +24,16 @@ export type TileProps = {
  */
 export const Tile = ({
                        title,
+                       titleClassName,
                        description,
+                       descriptionClassName,
                        onClick,
                        isSelected = false,
                        disabled = false,
                        prefix,
                        suffix,
                        normalClassName = 'hover:bg-primary/40 cursor-pointer',
-                       selectedClassName = ' bg-primary/20',
+                       selectedClassName = 'bg-primary/20',
                        disabledClassName = 'text-disabled-text bg-disabled-background cursor-not-allowed',
                        className
                      }: TileProps) => {
@@ -37,7 +42,7 @@ export const Tile = ({
       className={clsx(
         'flex-row-2 w-full items-center',
         {
-          [normalClassName]: !!onClick && !disabled,
+          [normalClassName]: onClick && !disabled,
           [selectedClassName]: isSelected && !disabled,
           [disabledClassName]: disabled,
         },
@@ -47,11 +52,26 @@ export const Tile = ({
     >
       {prefix}
       <div className="flex-col-0 w-full">
-        <h4 className={clsx(title.className ?? 'textstyle-title-normal')}>{title.value}</h4>
+        <span className={clsx(titleClassName ?? ('textstyle-title-normal'))}>{title}</span>
         {!!description &&
-          <span className={clsx(description.className ?? 'textstyle-description')}>{description.value}</span>}
+          <span className={clsx(descriptionClassName ?? 'textstyle-description')}>{description}</span>}
       </div>
       {suffix ?? (isSelected ? (<Check size={24}/>) : undefined)}
     </div>
+  )
+}
+
+export type ListTileProps = TileProps
+
+export const ListTile = ({
+  ...props
+}: ListTileProps) => {
+  return (
+    <Tile
+      {...props}
+      titleClassName={props.titleClassName ?? 'font-semibold'}
+      className={clsx('px-2 py-1 rounded-md', props.className)}
+      disabledClassName={props.disabledClassName ?? 'text-disabled-text cursor-not-allowed'}
+    />
   )
 }